[REFACTOR]MemberService와 GhostService분리작업 완료
