--- DontBeServer/src/main/java/com/dontbe/www/DontBeServer/api/member/dto/request/MemberNicknameCheckRequestDto.java
@@ -1,8 +1,0 @@
-package com.dontbe.www.DontBeServer.api.member.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record MemberNicknameCheckRequestDto (
-        @NotBlank String nickname
-){
-}
