+++ DontBeServer/src/main/java/com/dontbe/www/DontBeServer/api/member/service/MemberCommandService.java
@@ -1,0 +1,41 @@
+package com.dontbe.www.DontBeServer.api.member.service;
+
+import com.dontbe.www.DontBeServer.api.member.domain.Member;
+import com.dontbe.www.DontBeServer.api.member.dto.request.MemberProfilePatchRequestDto;
+import com.dontbe.www.DontBeServer.api.member.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional
+public class MemberCommandService {
+    private final MemberRepository memberRepository;
+
+    public void withdrawalMember(Long memberId) {
+        Member member = memberRepository.findMemberByIdOrThrow(memberId);
+        memberRepository.delete(member);
+    }
+
+    public void updateMemberProfile(Long memberId, MemberProfilePatchRequestDto memberProfilePatchRequestDto) {
+        Member existingMember = memberRepository.findMemberByIdOrThrow(memberId);
+
+        // 업데이트할 속성만 복사
+        if (memberProfilePatchRequestDto.nickname() != null) {
+            existingMember.updateNickname(memberProfilePatchRequestDto.nickname());
+        }
+        if (memberProfilePatchRequestDto.member_intro() != null) {
+            existingMember.updateMemberIntro(memberProfilePatchRequestDto.member_intro());
+        }
+        if (memberProfilePatchRequestDto.profile_url() != null) {
+            existingMember.updateProfileUrl(memberProfilePatchRequestDto.profile_url());
+        }
+        if (memberProfilePatchRequestDto.is_alarm_allowed() | !memberProfilePatchRequestDto.is_alarm_allowed()) {
+            existingMember.updateMemberIsAlarmAllowed(memberProfilePatchRequestDto.is_alarm_allowed());
+        }
+
+        // 저장
+        Member savedMember = memberRepository.save(existingMember);
+    }
+}
